norm clear.2
