+++ philo_bonus/src/philo_life_bonus.c
@@ -6,41 +6,13 @@
 /*   By: srakuma <srakuma@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 21:50:38 by srakuma           #+#    #+#             */
-/*   Updated: 2021/11/08 02:08:00 by srakuma          ###   ########.fr       */
+/*   Updated: 2021/11/08 03:46:54 by srakuma          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "philo_bonus.h"
-#include "ft_sem_sys_bonus.h"
 #include "philo_utils_bonus.h"
-
-long	read_cs(t_cs *cs)
-{
-	long	ret;
-
-	sem_wait(cs->sem);
-	ret = cs->val;
-	sem_post(cs->sem);
-	return (ret);
-}
-
-void	atomic_read_write(t_cs *cs, long val, t_op op)
-{
-	sem_wait(cs->sem);
-	if (op == ADD)
-		cs->val += val;
-	else if (op == SUB)
-		cs->val -= val;
-	else if (op == MUL)
-		cs->val *= val;
-	else if (op == DIV)
-		cs->val /= val;
-	else if (op == MOD)
-		cs->val %= val;
-	else
-		cs->val = val;
-	sem_post(cs->sem);
-}
+#include "read_write_cs.h"
+#include "philo_life.h"
 
 void	ft_print_philos_status(t_philo *philo, t_status status)
 {
@@ -61,17 +33,6 @@ void	ft_print_philos_status(t_philo *philo, t_status status)
 	sem_post(philo->all->for_print);
 }
 
-static void	ft_cleanup(t_philo *philo, sem_t *for_eaten_cnt)
-{
-	static long	eaten_cnt;
-
-	sem_post(philo->forks);
-	sem_post(philo->forks);
-	eaten_cnt++;
-	if (eaten_cnt == philo->all->min_times_eat)
-		sem_post(for_eaten_cnt);
-}
-
 static void	taking_forks(t_philo *philo)
 {
 	sem_wait(philo->reservation);
@@ -89,7 +50,8 @@ static void	taking_forks(t_philo *philo)
 
 void	ft_philo_eat(t_philo *philo, sem_t *for_eaten_cnt)
 {
-	long	lastmeal;
+	static long	eaten_cnt;
+	long		lastmeal;
 
 	if (philo->all->min_times_eat == 0)
 		exit (0);
@@ -99,54 +61,9 @@ void	ft_philo_eat(t_philo *philo, sem_t *for_eaten_cnt)
 	atomic_read_write(&philo->lastmeal, lastmeal, ASSIGN);
 	ft_msleep_until_time(lastmeal + philo->all->time_to_eat);
 	sem_post(philo->reservation);
-	ft_cleanup(philo, for_eaten_cnt);
-}
-
-void	*death_monitor(void *ph)
-{
-	long	death_time;
-	long	before_eaten_time;
-	long	after_eaten_time;
-	t_philo	*philo;
-
-	philo = (t_philo *)ph;
-	before_eaten_time = get_mtime();
-	after_eaten_time = before_eaten_time + 1;
-	atomic_read_write(&philo->lastmeal, before_eaten_time, ASSIGN);
-	while (before_eaten_time < after_eaten_time)
-	{
-		before_eaten_time = after_eaten_time;
-		death_time = read_cs(&philo->lastmeal) + philo->all->time_to_die;
-		ft_msleep_until_time(death_time);
-		after_eaten_time = read_cs(&philo->lastmeal);
-	}
-	exit(philo->philo_x);
-}
-
-void	the_life_of_philo(t_philo *philo)
-{
-	sem_t		*prepare;
-	sem_t		*term_sem;
-	pthread_t	tid;
-
-	prepare = sem_open(PREPARE, 0);
-	term_sem = sem_open(TERM_SEM, 0);
-	if (prepare == SEM_FAILED | term_sem == SEM_FAILED)
-	{
-		printf("philo %d\n", philo->philo_x);
-		perror(NULL);
-		exit(0);
-	}
-	sem_wait(term_sem);
-	sem_wait(prepare);
-	pthread_create(&tid, NULL, death_monitor, (void *)philo);
-	pthread_detach(tid);
-	ft_print_philos_status(philo, THINK);
-	while (1)
-	{
-		ft_philo_eat(philo, term_sem);
-		ft_print_philos_status(philo, SLEEP);
-		ft_msleep_until_time(get_mtime() + philo->all->time_to_sleep);
-		ft_print_philos_status(philo, THINK);
-	}
+	sem_post(philo->forks);
+	sem_post(philo->forks);
+	eaten_cnt++;
+	if (eaten_cnt == philo->all->min_times_eat)
+		sem_post(for_eaten_cnt);
 }
