+++ philo_bonus/src/start_philo_behavior.c
@@ -6,28 +6,12 @@
 /*   By: srakuma <srakuma@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 23:39:04 by srakuma           #+#    #+#             */
-/*   Updated: 2021/11/08 02:07:10 by srakuma          ###   ########.fr       */
+/*   Updated: 2021/11/08 03:36:58 by srakuma          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "philo_utils_bonus.h"
-#include "ft_sem_sys_bonus.h"
-#include "philo_bonus.h"
-
-void	kill_multi_child_proc(t_childlen *childlen, int sig)
-{
-	int	i;
-
-	i = 0;
-	while (i++ < childlen->size)
-	{
-		if (kill(childlen->pid[i - 1], sig) == EPERM)
-		{
-			ft_print_error_message("this parent process does not \
-					have a permission to send signal\n", __FILE__, __func__);
-		}
-	}
-}
+#include "start_philo_behavior.h"
+#include "using_thread_func.h"
 
 static bool	ft_start_child_process(t_philo *philo, t_childlen *childlen)
 {
@@ -49,13 +33,18 @@ static bool	ft_start_child_process(t_philo *philo, t_childlen *childlen)
 	return (true);
 }
 
-bool	init_childlen(t_childlen *childlen, t_philo *philo)
+static t_childlen	*init_childlen(t_philo *philo)
 {
+	t_childlen *childlen;
+
+	childlen = (t_childlen *)malloc(sizeof(t_childlen));
+	if (!childlen)
+		return (NULL);
 	childlen->pid = (pid_t *)malloc(sizeof(pid_t *) * philo->all->philo_num);
 	if (!childlen->pid)
 	{
-		ft_print_error_message("pid malloc error", __FILE__, __func__);
-		return (false);
+		free(childlen);
+		return (NULL);
 	}
 	memset(childlen->pid, 0, philo->all->philo_num);
 	sem_unlink(TERM_SEM);
@@ -63,14 +52,14 @@ bool	init_childlen(t_childlen *childlen, t_philo *philo)
 									700, philo->all->philo_num);
 	if (childlen->term_sem == SEM_FAILED)
 	{
+		free(childlen);
 		free(childlen->pid);
-		ft_print_error_message("sem_open error", __FILE__, __func__);
-		return (false);
+		return (NULL);
 	}
-	return (true);
+	return (childlen);
 }
 
-bool	synchronize_all_proc(t_philo *philo, sem_t *prepare, \
+static bool	synchronize_all_proc(t_philo *philo, sem_t *prepare, \
 												t_childlen *childlen)
 {
 	int	i;
@@ -87,43 +76,9 @@ bool	synchronize_all_proc(t_philo *philo, sem_t *prepare, \
 	return (true);
 }
 
-void	*wait_all_childlen_ate(void *val)
-{
-	int	j;
-	t_childlen	*childlen;
-
-	childlen = (t_childlen *)val;
-	j = 0;
-	while (j++ < childlen->size)
-		sem_wait(childlen->term_sem);
-	printf("in wait thread: time: %ld\n", get_mtime());
-	kill_multi_child_proc(childlen, SIGINT);
-	return (NULL);
-}
-
-void	wait_for_childlen(t_childlen *childlen)
-{
-	pthread_t	tid2;
-	int			status;
-
-	pthread_create(&tid2, NULL, wait_all_childlen_ate, (void *)childlen);
-	waitpid(-1, &status, 0);
-	if (WIFEXITED(status))
-	{
-		kill_multi_child_proc(childlen, SIGTERM);
-		if (WEXITSTATUS(status))
-			printf("%ld %d died\n", get_mtime(), _W_INT(status) >> 8);
-		pthread_detach(tid2);
-		free(childlen->pid);
-		return ;
-	}
-	pthread_join(tid2, NULL);
-	free(childlen->pid);
-}
-
 void	start_philo_life(t_philo *philo)
 {
-	t_childlen	childlen;
+	t_childlen	*childlen;
 	sem_t		*prepare;
 
 	sem_unlink(PREPARE);
@@ -133,12 +88,15 @@ void	start_philo_life(t_philo *philo)
 		ft_print_error_message("sem_open error", __FILE__, __func__);
 		return ;
 	}
-	if (init_childlen(&childlen, philo) == false)
-		return ;
-	if (synchronize_all_proc(philo, prepare, &childlen))
+	childlen = init_childlen(philo);
+	if (childlen == NULL)
+		return (ft_print_error_message("init t_childlen error", \
+											__FILE__, __func__));
+	if (synchronize_all_proc(philo, prepare, childlen) == false)
 	{
-		free(childlen.pid);
+		free(childlen->pid);
+		free(childlen);
 		return ;
 	}
-	return (wait_for_childlen(&childlen));
+	return (wait_for_childlen(childlen));
 }
