--- philo_bonus/test.txt
@@ -1,17 +1,0 @@
-1636302448782 1 is thinking
-1636302448782 1 has taken a fork
-1636302448782 1 has taken a fork
-1636302448782 1 is eating
-1636302448982 1 is sleeping
-1636302449182 1 is thinking
-1636302449197 1 died
-1636302448792 56 is thinking
-1636302448798 56 has taken a fork
-1636302448798 56 has taken a fork
-1636302448799 56 is eating
-1636302448999 56 is sleeping
-1636302448793 86 is thinking
-1636302448798 86 has taken a fork
-1636302448798 86 has taken a fork
-1636302448799 86 is eating
-1636302449000 86 is sleeping
